requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
pymysql==1.1.1
redis==6.4.0
paho-mqtt==2.1.0
